+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v1.2.0 (2021-02-20)
+
+* Adds a timeout for HTTP requests
+* Changes `Client.post` to `Client.make_http_request`
+
 ## v1.1.0 (2021-02-06)
 
 * Added a Gemfile that points to the Gemspec (closes #2)